+++ molecule/elasticsearch_test_modules/converge.yml
@@ -10,7 +10,7 @@
     elasticstack_full_stack: false
     elasticsearch_jna_workaround: true
     elasticsearch_disable_systemcallfilterchecks: true
-    elasticstack_release: 8
+    elasticstack_release: "{{ lookup('env', 'ELASTIC_RELEASE') | int}}"
     elasticsearch_heap: "1"
     elasticstack_no_log: false
   tasks:
@@ -46,7 +46,6 @@
         auth_user: elastic
         auth_pass: "{{ elasticstack_password.stdout }}"
         verify_certs: false
-        #ca_certs: /etc/elasticsearch/certs/http_ca.crt
 
     - name: Create elasticsearch user 'new-user'
       netways.elasticstack.elasticsearch_user:
@@ -63,4 +62,4 @@
         auth_user: elastic
         auth_pass: "{{ elasticstack_password.stdout }}"
         verify_certs: false
-        #ca_certs: /etc/elasticsearch/certs/http_ca.crt
+        ca_certs: /etc/elasticsearch/certs/http_ca.crt
