+++ molecule/elasticsearch_test_modules/molecule.yml
@@ -6,7 +6,7 @@ dependency:
 driver:
   name: docker
 platforms:
-  - name: elasticsearch_default1
+  - name: elasticsearch_default
     groups:
       - elasticsearch
     image: "geerlingguy/docker-${MOLECULE_DISTRO:-debian11}-ansible:latest"
@@ -16,16 +16,6 @@ platforms:
     cgroupns_mode: host
     privileged: true
     pre_build_image: true
-  #- name: elasticsearch_default2
-  #  groups:
-  #    - elasticsearch
-  #  image: "geerlingguy/docker-${MOLECULE_DISTRO:-debian11}-ansible:latest"
-  #  command: ${MOLECULE_DOCKER_COMMAND:-""}
-  #  volumes:
-  #    - /sys/fs/cgroup:/sys/fs/cgroup:rw
-  #  cgroupns_mode: host
-  #  privileged: true
-  #  pre_build_image: true
 provisioner:
   name: ansible
   env:
