+++ plugins/modules/elasticsearch_role.py
@@ -8,7 +8,7 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.netways.elasticstack.plugins.module_utils.api import (
+from ansible_collections.netways.elasticstack.plugins.module_utils.elasticsearch_role import (
     Role
 )
 
@@ -64,6 +64,13 @@ def run_module():
         failed=False,
         changed=False
     )
+
+    if module.params['state'] != 'absent' and module.params['state'] != 'present':
+        result['stderr'] = "Invalid state given. Please use 'absent' or 'present'"
+        result['failed'] = True
+        
+        module.exit_json(**result)
+
 
     role = Role(
         result=result,
