+++ plugins/modules/elasticsearch_user.py
@@ -8,7 +8,7 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.netways.elasticstack.plugins.module_utils.api import (
+from ansible_collections.netways.elasticstack.plugins.module_utils.elasticsearch_user import (
     User
 )
 
@@ -60,6 +60,13 @@ def run_module():
         changed=False
     )
 
+    if module.params['state'] != 'absent' and module.params['state'] != 'present':
+        result['stderr'] = "Invalid state given. Please use 'absent' or 'present'"
+        result['failed'] = True
+        
+        module.exit_json(**result)
+
+
     user = User(
         result=result,
         user_name=module.params['name'],
@@ -79,7 +86,6 @@ def run_module():
     result = user.return_result()
 
     module.exit_json(**result)
-
 
 if __name__ == "__main__":
     run_module()
